MAINT: Turn transformer's script to subpakcage
