+++ dtoolkit/transformer/factory.py
@@ -1,12 +1,7 @@
-from __future__ import annotations
-
-from typing import TYPE_CHECKING
+from typing import Callable
 
 from dtoolkit.transformer._util import snake_to_camel
 from dtoolkit.transformer.base import MethodTF
-
-if TYPE_CHECKING:
-    from typing import Callable
 
 
 def methodtf_factory(
