+++ dtoolkit/transformer/numpy/RavelTF.py
@@ -1,15 +1,10 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
-
 import numpy as np
 import pandas as pd
 
+from dtoolkit._typing import SeriesOrFrame
 from dtoolkit.transformer.base import NumpyTF
-
-
-if TYPE_CHECKING:
-    from dtoolkit._typing import SeriesOrFrame
 
 
 class RavelTF(NumpyTF):
