+++ dtoolkit/transformer/pipeline/FeatureUnion.py
@@ -1,0 +1,39 @@
+import pandas as pd
+from sklearn.pipeline import FeatureUnion as SKFeatureUnion
+
+from dtoolkit.transformer.base import Transformer
+
+
+class FeatureUnion(SKFeatureUnion, Transformer):
+    """
+    Concatenates results of multiple transformer objects.
+
+    See Also
+    --------
+    make_union
+        Convenience function for simplified feature union construction.
+
+    Notes
+    -----
+    Different to :obj:`sklearn.pipeline.FeatureUnion`.
+    This would let :obj:`~pandas.DataFrame` in and
+    :obj:`~pandas.DataFrame` out.
+
+    Examples
+    --------
+    >>> from dtoolkit.transformer import FeatureUnion
+    >>> from sklearn.decomposition import PCA, TruncatedSVD
+    >>> union = FeatureUnion([("pca", PCA(n_components=1)),
+    ...                       ("svd", TruncatedSVD(n_components=2))])
+    >>> X = [[0., 1., 3], [2., 2., 5]]
+    >>> union.fit_transform(X)
+    array([[ 1.5       ,  3.0...,  0.8...],
+           [-1.5       ,  5.7..., -0.4...]])
+    """
+
+    def _hstack(self, Xs):
+        if all(isinstance(i, (pd.Series, pd.DataFrame)) for i in Xs):
+            # merge all into one DataFrame and the index would use the common part
+            return pd.concat(Xs, axis=1, join="inner")
+
+        return super()._hstack(Xs)
