+++ dtoolkit/transformer/pipeline/Pipeline.py
@@ -1,11 +1,6 @@
-from __future__ import annotations
-
-import pandas as pd
 from pandas.util._decorators import doc
 from sklearn.base import clone
 from sklearn.pipeline import _fit_transform_one
-from sklearn.pipeline import _name_estimators
-from sklearn.pipeline import FeatureUnion as SKFeatureUnion
 from sklearn.pipeline import Pipeline as SKPipeline
 from sklearn.utils import _print_elapsed_time
 from sklearn.utils.metaestimators import available_if
@@ -14,7 +9,6 @@ from sklearn.utils.validation import check_memory
 from dtoolkit.transformer._util import transform_array_to_frame
 from dtoolkit.transformer._util import transform_frame_to_series
 from dtoolkit.transformer._util import transform_series_to_frame
-from dtoolkit.transformer.base import Transformer
 
 
 class Pipeline(SKPipeline):
@@ -118,103 +112,3 @@ class Pipeline(SKPipeline):
             Xt = transform_array_to_frame(transformer.inverse_transform(Xt), Xt)
 
         return transform_frame_to_series(Xt)
-
-
-def make_pipeline(
-    *steps: list[Transformer],
-    memory=None,
-    verbose: bool = False,
-) -> Pipeline:
-    """
-    Construct a :class:`Pipeline` from the given estimators.
-
-    This is a shorthand for the :class:`Pipeline` constructor; it does not
-    require, and does not permit, naming the estimators. Instead, their names
-    will be set to the lowercase of their types automatically.
-
-    See Also
-    --------
-    Pipeline : Class for creating a pipeline of transforms with a final
-        estimator.
-
-    Examples
-    --------
-    >>> from sklearn.naive_bayes import GaussianNB
-    >>> from sklearn.preprocessing import StandardScaler
-    >>> from dtoolkit.transformer import make_pipeline
-    >>> make_pipeline(StandardScaler(), GaussianNB(priors=None))
-    Pipeline(steps=[('standardscaler', StandardScaler()),
-                    ('gaussiannb', GaussianNB())])
-    """
-
-    return Pipeline(_name_estimators(steps), memory=memory, verbose=verbose)
-
-
-class FeatureUnion(SKFeatureUnion, Transformer):
-    """
-    Concatenates results of multiple transformer objects.
-
-    See Also
-    --------
-    make_union
-        Convenience function for simplified feature union construction.
-
-    Notes
-    -----
-    Different to :obj:`sklearn.pipeline.FeatureUnion`.
-    This would let :obj:`~pandas.DataFrame` in and
-    :obj:`~pandas.DataFrame` out.
-
-    Examples
-    --------
-    >>> from dtoolkit.transformer import FeatureUnion
-    >>> from sklearn.decomposition import PCA, TruncatedSVD
-    >>> union = FeatureUnion([("pca", PCA(n_components=1)),
-    ...                       ("svd", TruncatedSVD(n_components=2))])
-    >>> X = [[0., 1., 3], [2., 2., 5]]
-    >>> union.fit_transform(X)
-    array([[ 1.5       ,  3.0...,  0.8...],
-           [-1.5       ,  5.7..., -0.4...]])
-    """
-
-    def _hstack(self, Xs):
-        if all(isinstance(i, (pd.Series, pd.DataFrame)) for i in Xs):
-            # merge all into one DataFrame and the index would use the common part
-            return pd.concat(Xs, axis=1, join="inner")
-
-        return super()._hstack(Xs)
-
-
-def make_union(
-    *transformers: list[Transformer],
-    n_jobs: int = None,
-    verbose: bool = False,
-) -> FeatureUnion:
-    """
-    Construct a FeatureUnion from the given transformers.
-
-    See Also
-    --------
-    FeatureUnion
-        Class for concatenating the results of multiple transformer objects.
-
-    Notes
-    -----
-    Different to :obj:`sklearn.pipeline.make_union`.
-    This would let :obj:`~pandas.DataFrame` in and
-    :obj:`~pandas.DataFrame` out.
-
-    Examples
-    --------
-    >>> from sklearn.decomposition import PCA, TruncatedSVD
-    >>> from dtoolkit.transformer import make_union
-    >>> make_union(PCA(), TruncatedSVD())
-     FeatureUnion(transformer_list=[('pca', PCA()),
-                                   ('truncatedsvd', TruncatedSVD())])
-    """
-
-    return FeatureUnion(
-        _name_estimators(transformers),
-        n_jobs=n_jobs,
-        verbose=verbose,
-    )
